refactor(blankslate): rename from NonIdealState
