+++ packages/aksara-ui-core/src/components/non-ideal-state/components/Blankslate.test.tsx
@@ -1,10 +1,10 @@
 import * as React from 'react';
 import { render } from '@testing-library/react';
-import NonIdealState from './NonIdealState';
+import Blankslate from './Blankslate';
 
-describe('components/NonIdealState', () => {
+describe('components/Blankslate', () => {
   test('renders with title', () => {
-    const { getByText } = render(<NonIdealState title="Empty Text">Board test</NonIdealState>);
+    const { getByText } = render(<Blankslate title="Empty Text">Board test</Blankslate>);
 
     const title = getByText('Empty Text');
     expect(title).toBeInTheDocument();
